pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-scope-diff"
version = "0.1.0"
description = "Risk-focused diffs for agent manifests, MCP configs, and tool permission snapshots."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{ name = "agent-scope-diff contributors" }]
keywords = ["agents", "mcp", "security", "cli", "diff"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Quality Assurance",
]
dependencies = []

[project.urls]
Homepage = "https://github.com/pfrederiksen/agent-scope-diff"
Repository = "https://github.com/pfrederiksen/agent-scope-diff"
Issues = "https://github.com/pfrederiksen/agent-scope-diff/issues"

[project.optional-dependencies]
dev = ["PyYAML>=6.0", "pytest>=7.0"]

[project.scripts]
agent-scope-diff = "agent_scope_diff.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
